+++ spec/onlyoffice_webdriver_wrapper/web_driver_set_style_show_by_xpath_spec.rb
@@ -7,7 +7,7 @@ describe OnlyofficeWebdriverWrapper::WebDriver, '#set_style_show_by_xpath' do
 
   before do
     webdriver.open("file://#{Dir.pwd}/spec/html_examples/"\
-                      'style_helper.html')
+                      'set_style_show_by_xpath.html')
   end
 
   after { webdriver.quit }
@@ -32,6 +32,20 @@ describe OnlyofficeWebdriverWrapper::WebDriver, '#set_style_show_by_xpath' do
 
     it 'object is moved by y' do
       expect(webdriver.get_style_parameter('//span', 'top')).to eq('260px')
+    end
+  end
+
+  describe 'quotes in xpath' do
+    it '#set_style_show_by_xpath for xpath with single quotes' do
+      xpath = "//div[@id='foo']"
+      webdriver.set_style_show_by_xpath(xpath)
+      expect(webdriver.get_style_parameter(xpath, 'display')).to eq('block')
+    end
+
+    it '#set_style_show_by_xpath for xpath with double quotes' do
+      xpath = '//div[@id="foo"]'
+      webdriver.set_style_show_by_xpath(xpath)
+      expect(webdriver.get_style_parameter(xpath, 'display')).to eq('block')
     end
   end
 end