+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Extract `WebDriver#set_style_attribute` test to separate spec
 * Extract `WebDriver#set_style_show_by_xpath` test to separate spec
+* Simplify `WebDriver#set_style_show_by_xpath` by using `dom_element_by_xpath`
 
 ## 0.3.4 (2020-07-06)